#30 Fix compile errors for Motion Planning
The collection compiled fine with GCC but had issues with Clang-14.
Clang (all versions) current does not support templated alias type
deduction, so the UniformSampler space template parameter had to be
explicitly specified.

Additionally, Clang-14 had some weird bug when trying to use the
size_defined concept. Upgrading to Clang-16 resolved the issue. It's
unclear what the problem actually was, but Clang-16 is stable, so it
would be better to use the most recent compiler version. Not sure if it
was a bug with Clang or a problem with the code itself.

+++ motion_planning/examples/rrt.cpp
@@ -25,7 +25,9 @@ struct wheel::ResizeStrategy<Eigen::Vector2f> {
 
 template <>
 struct wheel::EuclideanDistance<Eigen::Vector2f> {
-    static auto distance(const Eigen::Vector2f &a, const Eigen::Vector2f &b) -> double { return (a - b).norm(); }
+    static auto distance(const Eigen::Vector2f &a, const Eigen::Vector2f &b) -> double {
+        return static_cast<double>((a - b).norm());
+    }
 };
 
 template <>
